requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
google-generativeai
langchain
langchain-community
faiss-cpu
langchain-google-genai
streamlit
requests
streamlit-option-menu
cohere
langchain-cohere
